requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --no-annotate --extra=dev --output-file=requirements-dev.txt pyproject.toml
annotated-types==0.6.0
anyio==4.3.0
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
contourpy==1.2.0
coverage==7.4.3
cryptography==42.0.5
cycler==0.12.1
db-dtypes==1.2.0
deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.8
filelock==3.13.1
fonttools==4.49.0
google-api-core==2.17.1
google-auth==2.28.2
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gotrue==2.4.1
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
identify==2.5.35
idna==3.6
iniconfig==2.0.0
kaleido==0.2.1
kiwisolver==1.4.5
matplotlib==3.8.3
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
pandas-gbq==0.23.1
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
postgrest==0.16.1
pre-commit==3.6.2
protobuf==4.25.3
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic-core==2.16.3
pydata-google-auth==1.8.2
pygithub==2.2.0
pyjwt==2.8.0
pynacl==1.5.0
pyparsing==3.1.2
pyproject-hooks==1.0.0
pytest==8.1.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
realtime==1.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
ruff==0.3.2
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
storage3==0.7.3
strenum==0.4.15
supabase==2.4.0
supafunc==0.3.3
tenacity==8.2.3
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
websockets==11.0.3
wrapt==1.16.0
